refactor : running button 추출
